Add redcap_version attribute to Project
On Project initialization PyCap will attempt to determine the REDCap version that the project is associated with. If it cannot it will warn the user that version information is unavailable. This implements [python-semanticversion](https://github.com/rbarrois/python-semanticversion) which allows for easy version comparison and assumes that REDCap will stick with its versioning conventions to date. If for whatever reason they differ, PyCap will simply return the string representation of the version from REDCap's API. Fixes #43.