:fire: Remove format parameter
`format` is not a valid parameter for this method, and it won't have any effect. But based on how the API is set up, specifically the RCRequest.validate() method, it expects format as one of the payload keys which is why we left it in. :thinking: I'll have to investigate if this is worth refactoring in the future