:recycle: Refactor conversion to csv for df imports
Similar to the prior change this was hard coded for record and metadata requests. But as this PR is demonstrating there a lot of other endpoints out there that allow imports and we certainly don't want to have to include each one in here. All that really seems to matter is whether or not there _is_ a named index. If there is, we want to keep it when we export. If there's not, then we don't need it.

As you can see from the test, importing from a dataframe will likely still require some finneagling on the part of the user to ensure all the formats are correct.