:bug: Backfill key fields for records export
In version 1.1.3, the record_id field would be included regardless of what combination of fields/forms were requested. The default REDCap API behavior is to NOT return the record_id field if it's not in the requested fields/forms. In version 2.0.0 we reverted to the default REDCap API behavior. This fixes that by backfilling not only the key fields such as "record_id" but also the "form_complete" fields as well, by passing the "fields" parameter explicitly in EVERY REDCap request