Add background_import param to import_records()
Setting this to True will ask REDCap to import records in the
background; this will work similarly to the background import in the
Data Import Tool.

Because this flag is not supported in old REDCap versions, if you do
not specify it, it will not be added to your request parameters at all
and you will get the default REDCap behavior.

Currently, REDCap defaults to foreground imports.